Make new GUI main
