+++ mnist_gui.py
@@ -1,21 +1,20 @@
-'''
-The basic framework of the GUI was ChatGPTed lol
-'''
-
-#imports
 import tkinter as tk
-import torch
-from torchvision import transforms
 import numpy as np
-import torch.nn as nn
-from typing import List
-from standard_cnn import get_prediction, ConvoNet
+from tkinter import *
+from PIL import Image, ImageDraw
+import io
+from torch import Tensor
+from torch.nn import AvgPool2d
+from standard_cnn import *
+from torchvision import transforms
 
-#constants
-GRID_SIZE = 28
-CELL_SIZE = 10
-BRIGHT_COLOR = 0.99
-DARK_COLOR = 0.7
+'''
+Prototyping app to create a better drawing brush for the real thing
+'''
+
+#color constants
+WHITE = 1
+BLACK = 0
 
 #load model
 MODEL = torch.load('mnist_classifier.pth')
@@ -26,69 +25,81 @@ transform: transforms.Compose = transforms.Compose([
         transforms.Normalize((0.1307,), (0.3081,))
         ])
 
-#tkinter app class
-class DrawingApp:
-    def __init__(self, root):
+class PaintApp:
+    
+    def __init__(self, root: tk.Tk, cell_size: int):
+        
         self.root = root
         
-        #canvas
-        self.canvas = tk.Canvas(root, width=GRID_SIZE*CELL_SIZE, height=GRID_SIZE*CELL_SIZE, bg='black')
-        self.canvas.pack()
-        self.canvas.bind('<B1-Motion>', self.draw)
+        self.CELL_SIZE = cell_size
+        self.WIDTH = int(cell_size * 2.5)
         
-        #data matrix
-        self.matrix = torch.Tensor(np.array([np.array([0 for _ in range(GRID_SIZE)]) for _ in range(GRID_SIZE)]))
+        self.canvas = tk.Canvas(self.root, 
+                                bg = 'black', 
+                                width = 28 * self.CELL_SIZE, 
+                                height = 28 * self.CELL_SIZE)
+        self.canvas.pack()
+        
+        self.clear_button = tk.Button(self.root, text='Clear', command=self.clear)
+        self.clear_button.pack()
+        
+        self.image: Image.Image = Image.new('I', (28 * self.CELL_SIZE, 28 * self.CELL_SIZE), BLACK)
+        self.drawer = ImageDraw.Draw(self.image)
+        
+        self.canvas.bind('<B1-Motion>', self.paint)
+        
+        self.save_button = tk.Button(self.root, text='Save', command=self.save)
+        self.save_button.pack()
+        
+        self.classify_button = tk.Button(self.root, text='Classify', command=self.classify)
+        self.classify_button.pack()
         
         #response label
         self.label = tk.Label(root, text='Click the Identify button to have your digit identified.')
         self.label.pack()
-
-    #draw method (main brush with tapered edges)
-    def draw(self, event):
-        x = event.x // CELL_SIZE
-        y = event.y // CELL_SIZE
-        half_width = 1
-        x1 = max(0, (x - half_width)) * CELL_SIZE
-        y1 = max(0, (y - half_width)) * CELL_SIZE
-        x2 = min(GRID_SIZE, (x + half_width + 1)) * CELL_SIZE
-        y2 = min(GRID_SIZE, (y + half_width + 1)) * CELL_SIZE
-        for x_i in range(max(0, (x - half_width)), min(GRID_SIZE, (x + half_width + 1))):
-            for y_i in range(max(0, (y - half_width)), min(GRID_SIZE, (y + half_width + 1))):
-                if (x_i, y_i) == (x, y):
-                    self.matrix[y_i][x_i] = BRIGHT_COLOR
-                elif not self.matrix[y_i][x_i]: #don't override previous 1s
-                    self.matrix[y_i][x_i] = DARK_COLOR
-        self.canvas.create_rectangle(x1, y1, x2, y2, fill='gray70', outline='gray70')
-        for x_i in range(max(0, (x - half_width)), min(GRID_SIZE, (x + half_width + 1))):
-            for y_i in range(max(0, (y - half_width)), min(GRID_SIZE, (y + half_width + 1))):
-                cell_x = x_i * CELL_SIZE
-                cell_y = y_i * CELL_SIZE
-                if self.matrix[y_i][x_i] == BRIGHT_COLOR:
-                    self.canvas.create_rectangle(cell_x, cell_y, 
-                                                 cell_x + CELL_SIZE, cell_y + CELL_SIZE, 
-                                                 fill='white', outline='white')
-            
-
+        
+        self.POOL = AvgPool2d((self.CELL_SIZE, self.CELL_SIZE), (self.CELL_SIZE, self.CELL_SIZE))
+        self.POOL.eval()
+    
+    def paint(self, event):
+        '''
+        Draw on screen (continuous circles)
+        '''
+        r = self.WIDTH
+        x_1, y_1 = event.x, event.y
+        x_2, y_2 = x_1 + r, y_1 + r
+        self.canvas.create_oval(x_1, y_1, x_2, y_2, fill='white', outline='white')
+        self.drawer.ellipse((x_1, y_1, x_2, y_2), fill=WHITE, outline=WHITE)
+    
     def clear(self):
-        self.canvas.delete('all')
-        self.matrix = torch.Tensor(np.array([np.array([0 for _ in range(GRID_SIZE)]) for _ in range(GRID_SIZE)]))
-        self.white_units = []
-            
-    def id(self):
-        image = self.matrix.unsqueeze(0).unsqueeze(0)
-        normed_image = transform(image)
+        '''
+        Clear drawing
+        '''
+        self.canvas.delete("all")
+        self.image: Image.Image = Image.new('I', (28 * self.CELL_SIZE, 28 * self.CELL_SIZE), BLACK)
+        self.drawer = ImageDraw.Draw(self.image)
+    
+    def save(self):
+        '''
+        Save image (testing only)
+        '''
+        self.image.save('./test.bmp')
+        
+    def classify(self):
+        '''
+        Classify number
+        '''
+        original_image = Tensor(np.asarray(self.image).copy())
+        downsampled_image = self.POOL(original_image.unsqueeze(0).unsqueeze(0))
+        normed_image = transform(downsampled_image)
+        self.label.config(text = f'Classifying...')
         prediction = get_prediction(MODEL.forward(normed_image)).item()
         self.label.config(text = f'Your number is {prediction}. Clear to try another!')
 
 if __name__ == '__main__':
-    root = tk.Tk()
-    root.title("Number Identification App")
-    app = DrawingApp(root)
-
-    clear_button = tk.Button(root, text='Clear Drawing', command=app.clear)
-    clear_button.pack(side='left')
     
-    id_button = tk.Button(root, text='Identify', command=app.id)
-    id_button.pack(side = 'left')
-
+    root = tk.Tk()
+    
+    PaintApp(root, 10)
+    
     root.mainloop()
